fixed naming issues in META.json files (estimator vs. counter)
This is a bit messy, because while the extensions use 'counter' naming
(adaptive_counter, bitmap_counter, ...), the distributions use 'estimator'
(adaptive_estimator). The META.json used the naming incorrectly, so it
was not possible to install the extensions using pgxn client.

The naming difference is still there, i.e. this does about the same thing

  $ pgxn --load adaptive_estimator -d testdb

  $ psql -d testdb -c "CREATE EXTENSION adaptive_counter"

A bit messy, the difference is pretty much unnecessary.